+++ IdempotenceKeys/PaymentsIdempotenceKey.cs
@@ -1,8 +1,8 @@
 namespace PixHub.IdempotenceKeys;
 
-public class PaymentsIdempotenceKey(string cpf, int amount, int pixKeyId)
+public class PaymentsIdempotenceKey(string cpf, int amount, long pixKeyId)
 {
   public string Cpf { get; } = cpf;
   public int Amount { get; } = amount;
-  public int PixKeyId { get; } = pixKeyId;
+  public long PixKeyId { get; } = pixKeyId;
 }
