+++ Models/PaymentProviderAccount.cs
@@ -3,11 +3,11 @@ using System.ComponentModel.DataAnnotations.Schema;
 
 namespace PixHub.Models;
 
-public class PaymentProviderAccount(string accountNumber, string agency, int userId, int paymentProviderId)
+public class PaymentProviderAccount(string accountNumber, string agency, long userId, long paymentProviderId)
 {
   [Key]
   [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-  public int Id { get; set; }
+  public long Id { get; set; }
   public string AccountNumber { get; set; } = accountNumber;
   public string Agency { get; set; } = agency;
 
@@ -15,9 +15,9 @@ public class PaymentProviderAccount(string accountNumber, string agency, int use
   public ICollection<Payments>? Payments { get; set; }
 
   public User? User { get; set; }
-  public int UserId { get; set; } = userId;
+  public long UserId { get; set; } = userId;
   public PaymentProvider? PaymentProvider { get; set; }
-  public int PaymentProviderId { get; set; } = paymentProviderId;
+  public long PaymentProviderId { get; set; } = paymentProviderId;
 
   public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
   public DateTime UpdatedAt { get; set; } = DateTime.UtcNow;
